requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
Django==4.2.2
django-bootstrap-v5==1.0.11
django-cors-headers==4.3.1
django-countries==7.5.1
django-rest-framework==0.1.0
django-seed==0.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
Faker==25.2.0
mysqlclient==2.2.0
Pillow==10.0.1
psycopg2==2.9.9
pygame==2.5.2
Pygments==2.18.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
toposort==1.10
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.2.1
